Promene u filtriranju farmaceuta

+++ ISAISA/src/main/java/com/example/ISAISA/DTO/FilterEmployeesDTO.java
@@ -2,31 +2,31 @@ package com.example.ISAISA.DTO;
 
 public class FilterEmployeesDTO {
 
-    private Integer ratingOver;
-    private Integer ratingUnder;
+    private Float ratingOver;
+    private Float ratingUnder;
     private String pharmacyName;
 
     public FilterEmployeesDTO() {
     }
 
-    public FilterEmployeesDTO(Integer ratingOver, Integer ratingUnder, String pharmacyName) {
+    public FilterEmployeesDTO(Float ratingOver, Float ratingUnder, String pharmacyName) {
         this.ratingOver = ratingOver;
         this.ratingUnder = ratingUnder;
         this.pharmacyName = pharmacyName;
     }
 
-    public FilterEmployeesDTO(Integer ratingOver, Integer ratingUnder) {
+    public FilterEmployeesDTO(Float ratingOver, Float ratingUnder) {
         this.ratingOver = ratingOver;
         this.ratingUnder = ratingUnder;
     }
 
-    public Integer getRatingOver() { return ratingOver; }
+    public Float getRatingOver() { return ratingOver; }
 
-    public void setRatingOver(Integer ratingOver) { this.ratingOver = ratingOver; }
+    public void setRatingOver(Float ratingOver) { this.ratingOver = ratingOver; }
 
-    public Integer getRatingUnder() { return ratingUnder; }
+    public Float getRatingUnder() { return ratingUnder; }
 
-    public void setRatingUnder(Integer ratingUnder) { this.ratingUnder = ratingUnder; }
+    public void setRatingUnder(Float ratingUnder) { this.ratingUnder = ratingUnder; }
 
     public String getPharmacyName() { return pharmacyName; }
